Improvements to questions
