+++ app/documentation/questions_redis/index.js
@@ -12,6 +12,8 @@ const cache = require("helper/express-disk-cache")(config.cache_directory);
 const flushOptions = { host: config.host, path: "/https/temporary/questions" };
 const flush = () => cache.flush(flushOptions);
 const questions = require("models/questions");
+const render = require('./render');
+const Paginator = require("./paginator");
 
 Questions.use(["/ask", "/:id/edit", "/:id/new"], require("dashboard/session"));
 Questions.use(["/ask", "/:id/edit", "/:id/new"], urlencoded);
@@ -412,106 +414,12 @@ Questions.get(
   }
 );
 
-function Paginator(page, itemsPerPage, totalItems, base) {
-  // Data for pagination
-  let pages_count = Math.ceil(totalItems / itemsPerPage);
-
-  // Paginator object for the view
-  let paginator = {};
-
-  // total pages
-  let next_page = false;
-  let previous_page = false;
-
-  if (page < pages_count) next_page = page + 1; // next page value only if current page is not last
-  if (page > 1) previous_page = page - 1; // next page value only if current page is not last
-
-  if (pages_count > 1) {
-    // create paginator only if there are more than 1 pages
-    paginator = {
-      pages: [], // array of pages [{page: 1, current: true}, {...}, ... ]
-      next_page: next_page, // next page int
-      previous_page: previous_page, // next page int
-      topics_count: totalItems, // total number of topics
-    };
-
-    // Produces pagination links like this:
-    // 1 ... 56 57 [58] 59 60 ... 90
-
-    const number_of_links_each_side_of_current_page = 2;
-
-    for (let i = 1; i <= pages_count; i++) {
-      // filling pages array
-      if (
-        i === 1 ||
-        i === page ||
-        i === pages_count ||
-        (i < number_of_links_each_side_of_current_page * 2 + 1 &&
-          page < number_of_links_each_side_of_current_page * 2 + 1) ||
-        Math.abs(page - i) < number_of_links_each_side_of_current_page + 1
-      ) {
-        paginator.pages.push({ page: i, current: i === page, base });
-      } else if (
-        Math.abs(page - i) === number_of_links_each_side_of_current_page + 1 ||
-        (i === number_of_links_each_side_of_current_page * 2 + 1 &&
-          page < number_of_links_each_side_of_current_page * 2 + 1)
-      ) {
-        paginator.pages.push({ elipsis: true });
-      }
-    }
-  }
-
-  console.log("Paginator", paginator);
-  return paginator;
-}
-
-const he = require("he");
-const hljs = require("highlight.js");
-const cheerio = require("cheerio");
-const { urlencoded } = require("body-parser");
-
 // Removes everything forbidden by XML 1.0 specifications,
 // plus the unicode replacement character U+FFFD
 function removeXMLInvalidChars(string) {
-  var regex =
-    /((?:[\0-\x08\x0B\f\x0E-\x1F\uFFFD\uFFFE\uFFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF]))/g;
-  return string.replace(regex, "");
-}
-
-function render(input) {
-  let html = input;
-
-  try {
-    html = highlight(marked(input));
-  } catch (e) {
-    console.error(e);
+    var regex =
+      /((?:[\0-\x08\x0B\f\x0E-\x1F\uFFFD\uFFFE\uFFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF]))/g;
+    return string.replace(regex, "");
   }
-
-  return html;
-}
-
-function highlight(html) {
-  const $ = cheerio.load(html);
-
-  $("pre code").each(function () {
-    try {
-      var lang = $(this).attr("class").split("language-")[1];
-      console.log("lang:", lang);
-      if (!lang) return;
-      var code = $(this).text();
-      code = he.decode(code);
-
-      // For some reason highlight
-      // doesn't play nicely with already-decoded
-      // apostrophes like ' &#84; etc...
-
-      var highlighted = hljs.highlight(lang, code).value;
-
-      $(this).html(highlighted).addClass("hljs").addClass(lang);
-    } catch (e) {}
-  });
-
-  return $.html();
-}
 
 module.exports = Questions;
